+++ tree/Cargo.toml
@@ -16,7 +16,7 @@ crate-type = ["lib", "cdylib"]
 
 [dependencies]
 itertools = "0.10"
-mpi = {version = "0.6.*", optional = True}
+mpi = {version = "0.6.*", optional = true}
 memoffset = "0.6"
 rand = "0.8.*"
 hyksort = { path = "../hyksort"}
@@ -24,3 +24,5 @@ vtkio = "0.6"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 
+[features]
+mpi = ["dep:mpi"]
