+++ tree/src/implementations/impl_point.rs
@@ -1,0 +1,32 @@
+use std::cmp::Ordering;
+use std::hash::{Hash, Hasher};
+
+use crate::types::point::Point;
+
+impl PartialEq for Point {
+    fn eq(&self, other: &Self) -> bool {
+        self.key == other.key
+    }
+}
+
+impl Eq for Point {}
+
+impl Ord for Point {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.key.cmp(&other.key)
+    }
+}
+
+impl PartialOrd for Point {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        // less_than(&self.morton, &other.morton)
+        Some(self.key.cmp(&other.key))
+    }
+}
+
+impl Hash for Point {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        self.key.hash(state);
+    }
+}
+
