+++ tree/src/types/mod.rs
@@ -1,4 +1,7 @@
-//! # Required types and their MPI equivalences.
+//! # Type declaration
 pub mod domain;
 pub mod morton;
 pub mod point;
+pub mod single_node;
+#[cfg(feature = "mpi")]
+pub mod  multi_node;
